Implemented Resources.
Resources are static files which have to be served by FuzeWorks. The developer can register a directory in which static files can be found, which shall be loaded if the folder is requested.
This system should be avoided for high-performance applications. It is recommended to make special configurations in the web server in those kinds of cases.